Remove single-use utility functions
Inline isValidEmailAddress, formatSizeHuman, and isElementWithin to reduce utility file count
