--- components/isElementWithin.ts
@@ -1,2 +1,0 @@
-export const isElementWithin = (target: HTMLElement, els: (HTMLElement | null)[]) =>
-  els.some((el) => el?.contains(target));
