--- components/utils/email.ts
@@ -1,1 +1,0 @@
-export const isValidEmailAddress = (email: string) => /^[^\s@]+@[^\s@]+\.[^\s@]+$/u.test(email);
